move code to new module for validation
