+++ utils/wasm-builder/src/optimize.rs
@@ -1,3 +1,21 @@
+// This file is part of Gear.
+
+// Copyright (C) 2022-2024 Gear Technologies Inc.
+// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+
 use crate::stack_end;
 use anyhow::{Context, Result};
 #[cfg(not(feature = "wasm-opt"))]
@@ -68,8 +86,8 @@ impl Optimizer {
 
     /// Strips all custom sections.
     ///
-    /// Presently all custom sections are not required so they can be stripped safely.
-    /// The name section is already stripped by `wasm-opt`.
+    /// Presently all custom sections are not required so they can be stripped
+    /// safely. The name section is already stripped by `wasm-opt`.
     pub fn strip_custom_sections(&mut self) {
         self.module
             .sections_mut()
@@ -126,8 +144,8 @@ pub struct OptimizationResult {
 
 /// Attempts to perform optional Wasm optimization using `binaryen`.
 ///
-/// The intention is to reduce the size of bloated Wasm binaries as a result of missing
-/// optimizations (or bugs?) between Rust and Wasm.
+/// The intention is to reduce the size of bloated Wasm binaries as a result of
+/// missing optimizations (or bugs?) between Rust and Wasm.
 pub fn optimize_wasm(
     source: PathBuf,
     destination: PathBuf,
